Fix several warnings and build issues.

+++ src/ox/std/types.hpp
@@ -7,14 +7,14 @@
  */
 #pragma once
 
-typedef char               int8_t;
+typedef signed char        int8_t;
 typedef unsigned char      uint8_t;
 typedef short              int16_t;
 typedef unsigned short     uint16_t;
 typedef int                int32_t;
 typedef unsigned int       uint32_t;
 typedef unsigned           uint_t;
-typedef long long          int64_t;
+typedef long               int64_t;
 typedef unsigned long long uint64_t;
 
 namespace ox {
